+++ Assets/Scripts/Interfaces/DamageInterface.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface iDamageable
 {
-    public void receiveDamage(int amount);
+    public void receiveDamage(float amount);
 }
